Deferred-ledger sweep: move FileUtilities/ContextMenuUtilities, prune stale rows
Re-read every section 7 row against the current tree. Three rows named
blockers that had since landed:

- Utilities/FileUtilities.cs (blocked on AutomationProfile.Current, which
  landed in Core in 6a).
- Utilities/ContextMenuUtilities.cs (blocked on AutomationProfile.Current,
  FileUtilities.GetExePath(), and the IoUtilities split - all resolved).
- Utilities/FileAssociationUtilities.cs (blocked on FileUtilities.GetExePath()).

ContextMenuUtilities.cs moved unsplit to Core.Windows.

FileUtilities.cs needed one split: GetOpenDocumentFilter() also calls
GrabFrameFileUtilities (.GrabFrameFileExtension, .GetGrabFrameFileFilter()),
which stays app-side - blocked on HistoryInfo per its own section 7 row,
untouched here since Services/HistoryService.cs is explicitly out of scope
for this sweep. Everything else in FileUtilities (12 other members, a
dozen-plus call sites across the app) moved to Core.Windows keeping the
name. GetOpenDocumentFilter() alone (3 call sites: App.xaml.cs,
EditTextWindow.xaml.cs, one test) moved into a new app-side
Text-Grab/Utilities/OpenDocumentFilterUtilities.cs, calling back into two
of FileUtilities's helpers (GetVisualDocumentFilterPattern,
GetExtensionsFilterPattern) widened from private to internal for exactly
that caller. AppUtilities.IsPackaged() calls in FileUtilities became
PackageIdentity.IsPackaged() - the established 4e substitution, since
AppUtilities lives in the app and Core.Windows cannot see it.

FileAssociationUtilities.cs did not move: its GrabFrameExtensionKeyPath
constant references GrabFrameFileUtilities.GrabFrameFileExtension
directly - the same HistoryInfo blocker one level removed, not previously
recorded. Its section 7 row was rewritten with the real blocker rather
than forcing the move.

Utilities/TesseractHelper.cs's row was stale - it moved to Core.Windows in
batch 4b and the row was simply never removed - deleted.

Self-assembly-reflection audit: no GetExecutingAssembly/GetEntryAssembly/
GetCallingAssembly, typeof(x).Assembly, Assembly.GetName/.Location/
.CodeBase, AssemblyMetadata, AppContext.BaseDirectory or manifest-resource
lookups in either moved file. FileUtilities.GetExePath() already used
Environment.ProcessPath rather than assembly-location reflection, so it
needed no change for this.

Gate: dotnet build Tests/Tests.csproj -c Debug -p:Platform=x64 - clean.
Tests.Core 514, Tests.Core.Windows 3, Tests 993 passed/7 skipped - 1510
total, conserved.
